fix(sync): diff and commit URL outputs
